Merge
From: Maarten ter Huurne <maarten@treewalker.org>
Date: Thu, 16 May 2013 05:00:43 +0200
Subject: [PATCH] net: wireless: rtl8192cu: Fixed usage of kthread_run

The returned value is a pointer to a struct, not an integer.

This fixes driver init on MIPS, where pointers are likely to be
negative when interpreted as an integer (since they are allocated
in kseg0).